refactor(agent-core): split the Zod converter by responsibility
The converter crossed the 300-line ceiling as the review findings landed. Split
one way, conversion -> inspection:
`packages/agent-core/src/schema/zod-schema-inspect.ts` answers questions about a
Zod schema (what does it wrap, is this field required, which unknown-key mode,
what enum values), and `packages/agent-core/src/schema/zod-to-json-schema.ts`
decides what subset node to emit. The dependency runs in one direction only, so
the mutual recursion between the object walk and the node walk stays in a single
file rather than becoming a cycle across two.

`extractEnumValues`, `hasValidationConstraints` and `getSchemaTypeName` move with
the inspection half and are re-exported from the package barrel unchanged, so no
consumer import changes.
