+++ packages/agent-core/src/schema/zod-to-json-schema.test.ts
@@ -1,11 +1,11 @@
 import { describe, it, expect } from 'vitest';
 import type { IZodSchema, IZodSchemaDef } from './zod-schema-types';
 import {
-  zodToJsonSchema,
   extractEnumValues,
-  hasValidationConstraints,
   getSchemaTypeName,
-} from './zod-to-json-schema';
+  hasValidationConstraints,
+} from './zod-schema-inspect';
+import { zodToJsonSchema } from './zod-to-json-schema';
 
 /**
  * Create a mock Zod schema definition.
